requirements.txt
fastapi
uvicorn[standard]
gunicorn
pydantic
pandas
numpy
scipy
google-cloud-firestore
google-cloud-storage
google-cloud-aiplatform